backend/requirements-viz.txt
# Dependencias para visualización 3D del RAG
numpy>=1.24.0
scikit-learn>=1.3.0
umap-learn>=0.5.4

